requirements.txt
fastapi[standard]
uvicorn[standard]
passlib[bcrypt]
pymongo
python-dotenv
# bson
PyJWT